requirements.txt
streamlit
pandas
plotly-express
matplotlib
seaborn
numpy
scikit-learn